+++ exchangelib/services/upload_items.py
@@ -1,13 +1,10 @@
-from .common import EWSAccountService
+from .common import EWSAccountService, to_item_id
 from ..properties import ItemId, ParentFolderId
 from ..util import create_element, set_xml_value, add_xml_child, MNS
 
 
 class UploadItems(EWSAccountService):
     """MSDN: https://docs.microsoft.com/en-us/exchange/client-developer/web-service-reference/uploaditems-operation
-
-    TODO: This currently has the existing limitation of only being able to upload items that do not yet exist in the
-     database. The full spec also allows actions "Update" and "UpdateOrCreate".
     """
 
     SERVICE_NAME = 'UploadItems'
@@ -21,7 +18,8 @@ class UploadItems(EWSAccountService):
         """Upload given items to given account.
 
         'items' is an iterable of tuples where the first element is a Folder instance representing the ParentFolder
-        that the item will be placed in and the second element is a Data string returned from an ExportItems.
+        that the item will be placed in and the second element is a tuple containing an optional ItemId, an optional
+        Item.is_associated boolean, and a Data string returned from an ExportItems.
         call.
 
         :param items:
@@ -29,10 +27,16 @@ class UploadItems(EWSAccountService):
         uploaditems = create_element('m:%s' % self.SERVICE_NAME)
         itemselement = create_element('m:Items')
         uploaditems.append(itemselement)
-        for parent_folder, data_str in items:
-            item = create_element('t:Item', attrs=dict(CreateAction='CreateNew'))
+        for parent_folder, (item_id, is_associated, data_str) in items:
+            # TODO: The full spec also allows the "UpdateOrCreate" create action.
+            item = create_element('t:Item', attrs=dict(CreateAction='Update' if item_id else 'CreateNew'))
+            if is_associated is not None:
+                item.set('IsAssociated', 'true' if is_associated else 'false')
             parentfolderid = ParentFolderId(parent_folder.id, parent_folder.changekey)
             set_xml_value(item, parentfolderid, version=self.account.version)
+            if item_id:
+                itemid = to_item_id(item_id, ItemId, version=self.account.version)
+                set_xml_value(item, itemid, version=self.account.version)
             add_xml_child(item, 't:Data', data_str)
             itemselement.append(item)
         return uploaditems